feat: move UpdateClassifier to its own file improve documentation
